+++ konduit-serving-data/konduit-serving-image/src/main/resources/META-INF/services/ai.konduit.serving.pipeline.api.step.PipelineStepRunnerFactory
@@ -22,6 +22,6 @@ ai.konduit.serving.data.image.step.grid.draw.DrawGridStepRunnerFactory
 ai.konduit.serving.data.image.step.segmentation.index.DrawSegmentationStepRunnerFactory
 ai.konduit.serving.data.image.step.grid.crop.CropGridStepRunnerFactory
 ai.konduit.serving.data.image.step.face.DrawFaceKeyPointsStepRunnerFactory
-ai.konduit.serving.data.image.step.point.convert.ConvertBoundingBoxToPointStepRunnerFactory
+ai.konduit.serving.data.image.step.bb.point.BoundingBoxToPointStepRunnerFactory
 ai.konduit.serving.data.image.step.point.perspective.convert.PerspectiveTransformStepRunnerFactory
 ai.konduit.serving.data.image.step.resize.ImageResizeFactory
